# docs/swagger.yaml
definitions:
...
  api.Payment:
    properties:
      amount:
        type: number
      currency:
        type: string
      date:
        type: string
      debtMode:
        type: string
      debtors:
        items:
          type: string
        type: array
      description:
        type: string
      payer:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
  api.PaymentResponse:
    properties:
      amount:
        type: number
      currency:
        type: string
      date:
        type: string
      debtors:
        items:
          $ref: '#/definitions/api.UserSummary'
        type: array
      description:
        type: string
      id:
        type: integer
      payer:
        $ref: '#/definitions/api.UserSummary'
      tags:
        items:
          type: string
        type: array
    type: object
...
  api.UserStatsResponse:
    properties:
      currency:
        type: string
      paymentsIncluded:
        type: integer
      spentOut:
        type: number
      totalSpent:
        type: number
      user:
        $ref: '#/definitions/api.UserSummary'
    type: object
  api.UserSummary:
    properties:
      discordHandle:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/payment:
    post:
      consumes:
      - application/json
      parameters:
      - description: Payment payload
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/api.Payment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new payment
      tags:
      - payment
  /api/v1/payment/{id}:
    delete:
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a payment by ID
      tags:
      - payment
    get:
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PaymentResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a payment by ID
      tags:
      - payment
    put:
      consumes:
      - application/json
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payment payload
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/api.Payment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a payment by ID
      tags:
      - payment
  /api/v1/payment/all:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.PaymentResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all payments
      tags:
      - payment
  /api/v1/payment/tag/{tag}:
    get:
      parameters:
      - description: Payment tag
        in: path
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.PaymentResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get payments by tag
      tags:
      - payment
  /api/v1/payment/tags:
    get:
      parameters:
      - collectionFormat: multi
        description: Payment tags
        in: query
        items:
          type: string
        name: tags
        required: true
        type: array
      - default: and
        description: 'Tag operation: and/or'
        enum:
        - and
        - or
        in: query
        name: op
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.PaymentResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get payments by tags
      tags:
      - payment
  /api/v1/stats/user:
    get:
      parameters:
      - description: User name
        in: query
        name: user
        required: true
        type: string
      - default: USD
        description: Currency for generated stats
        in: query
        name: currency
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserStatsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user spending statistics
      tags:
      - stats
...
swagger: "2.0"